Add files via upload
In current scenario MS Paint or different paint software’s are used for editing or drawing images on screen with traditional pointing devices are used. For editing image or drawing shape user do not have any other option except keyboard ,mouse to use paint software. So to avoid this we are proposing a Smart paint software which can be used without any hardware interface.